// src/app/models/production.ts
import { BaseModel } from './base-model';
import { ProductionProcess } from './production-process';
import { ProductionProductInfo } from './production-product-info';
import { ProductionHistory } from './production-history';

export class Production extends BaseModel
{
    id: number;
    orderId: string;
    orderName: string;
    workOrderNo: string;
    lineItemId: number;
    vendorId: string;
    vendorName: string;
    customerId: string;
    customerName: string;
    batch: string;
    action: string;
    name: string;
    description: string;
    due: string;
    decoTypeId: number;
    decoTypeName: string;
    detailName: string;
    detailCount: number;
    qty: number;
    position: string;
    detailDescription: string;
    machineId: number;
    machineName: string;
    hours: number;
    productId: number;
    productName: string;
    statusId: number;
    statusName: string;
    priorityId: number;
    priorityName: string;
    decoRecordId: number;
    productionRate: number;
    processes: ProductionProcess[];
    variationId: string;
    variationName: string;
    designId: number;
    designNum: string;
    artworkId: number;
    batchJob: boolean;
    batchMaster: boolean;
    completedBatch: boolean;
    productInfo: ProductionProductInfo[];
    batchJobs: Production[];
    jobHistory: ProductionHistory[];
    size: string;
    color: string;
    pinned: boolean;
    variation: any[];

    setData(production) {
        this.setFloat(production, 'id');
        this.setString(production, 'orderId');
        this.setString(production, 'orderName');
        this.setString(production, 'workOrderNo');
        this.setString(production, 'lineItemId');
        this.setString(production, 'vendorId');
        this.setString(production, 'vendorName');
        this.setString(production, 'customerId');
        this.setString(production, 'customerName');
        this.setString(production, 'batch');
        this.setString(production, 'action');
        this.setString(production, 'name');
        this.setString(production, 'description');
        this.setString(production, 'due');
        this.setFloat(production, 'decoTypeId');
        this.setString(production, 'decoTypeName');
        this.setString(production, 'detailName');
        this.setFloat(production, 'detailCount');
        this.setFloat(production, 'qty');
        this.setString(production, 'position');
        this.setString(production, 'detailDescription');
        this.setFloat(production, 'machineId');
        this.setString(production, 'machineName');
        this.setFloat(production, 'hours');
        this.setFloat(production, 'productId');
        this.setString(production, 'productName');
        this.setFloat(production, 'statusId');
        this.setString(production, 'statusName');
        this.setFloat(production, 'priorityId');
        this.setString(production, 'priorityName');
        this.setFloat(production, 'decoRecordId');
        this.setFloat(production, 'productionRate');
        this.setObjectArray(production.processes || [], 'processes', ProductionProcess);
        this.setString(production, 'variationId');
        this.setString(production, 'variationName');
        this.setFloat(production, 'designId');
        this.setString(production, 'designNum');
        this.setFloat(production, 'artworkId');
        this.setBoolean(production, 'batchJob');
        this.setBoolean(production, 'batchMaster');
        this.setBoolean(production, 'completedBatch');
        this.setObjectArray(production.productInfo || [], 'productInfo', ProductionProductInfo);
        this.setObjectArray(production.batchJobs || [], 'batchJobs', Production);
        this.setObjectArray(production.history || [], 'jobHistory', ProductionHistory)
        this.setString(production, 'size');
        this.setString(production, 'color');
        this.setBoolean(production, 'pinned');
        this.setObjectArray(production.variation || [], 'variation');
    }
}